feat: working on xml2ass

+++ crates/nicolive/src/model.rs
@@ -177,7 +177,7 @@ pub struct DanmakuMessageChat {
     pub user_id: String,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub premium: Option<u8>,
+    pub premium: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub anonymity: Option<u8>,
